pyverdict should be tested
