--- core-scala/src/main/scala/edu/umich/verdict/App.scala
@@ -1,15 +1,0 @@
-package edu.umich.verdict
-
-/**
- * @author ${user.name}
- */
-object App {
-  
-  def foo(x : Array[String]) = x.foldLeft("")((a,b) => a + b)
-  
-  def main(args : Array[String]) {
-    println( "Hello World!" )
-    println("concat arguments = " + foo(args))
-  }
-
-}
