+++ python/pyverdict/__init__.py
@@ -2,5 +2,5 @@
 from context import VerdictSparkContext
 
 __all__ = [
-    "VerdictSparkContext"
+    "VerdictHiveContext"
 ]
